Create ConsultantEvaluation.php
This takes the information from the ConsultantForm and uses it to add to the database. If you are trying to test this, you will have to move both the ConsultantForm.html and ConsultantEvaluation.php to C:\xampp\htdocs somewhere. Additionally, you will have to create the writing center database. You can use the WritingCenter.sql script to create the database. This does not create any tuples.